move operator bootstrappers to `pkg`
