+++ pkg/operator/bootstrappers/verify_version.go
